move AsterixCharacter, AsterixCharacterRepo and AsterixController to equivalent packages
