+++ src/main/java/com/github/ingbeck/asterixapi/repository/AsterixCharacterRepo.java
@@ -1,5 +1,6 @@
-package com.github.ingbeck.asterixapi;
+package com.github.ingbeck.asterixapi.repository;
 
+import com.github.ingbeck.asterixapi.model.AsterixCharacter;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface AsterixCharacterRepo extends MongoRepository<AsterixCharacter, String> {
